Skip plan tests based on runtime_data
The stage Mozilla accounts / FxA server has the US monthly and yearly
plans setup, but not the ones for other countries supported in
production. This reads /api/v1/runtime_data, and allows skipping tests
if we're outside of US / CA and on stage, or if the current country is
not supported at all.